Add support for custom importers
Importers are declared by the host in `CompileRequest` using opaque
IDs, which the compiler then refers to when resolving imports. It adds
two outbound requests: one to convert a user-authored URL to a
canonical URL, and one to load the stylesheet for the canonical URL.
The canonicalization step is necessary for implementing the load-once
behavior in the new Sass module system.

Closes #4